+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+* Regenerate session ID on login / logout 
+
 ### v0.2.0 (2018-02-20)
 
 * Bump kohana-extras dependency to 0.2